+++ Stitch/Graph/StitchAI/GraphPrompting/Model/SchemaVersions/V0/AICodeGen/AICodeGenRequestBody_V0.swift
@@ -7,57 +7,6 @@
 
 import Foundation
 
-// TODO: move
-protocol StitchAIRequestableFunctionBody: Encodable {
-    var tools: [OpenAIFunction] { get }
-    var tool_choice: OpenAIFunction { get }
-}
-
-extension StitchAIRequestableFunctionBody {
-    var functionName: String {
-        self.tool_choice.function.name
-    }
-}
-
-enum AICodeEditBody_V0 {
-    // https://platform.openai.com/docs/api-reference/making-requests
-    struct AICodeEditRequestBody: StitchAIRequestableFunctionBody {
-        static let markdownLocation = "AICodeGenSystemPrompt_V0"
-        
-        let model: String = "o4-mini-2025-04-16"
-        let n: Int = 1
-        let temperature: Double = 1.0
-        let messages: [OpenAIMessage]
-        let tools = StitchAIRequestBuilder_V0.StitchAIRequestBuilderFunctions.allFunctions
-        let tool_choice = StitchAIRequestBuilder_V0.StitchAIRequestBuilderFunctions.codeEditor.function
-        let stream: Bool = false
-        
-        init(userPrompt: String,
-             toolMessages: [OpenAIMessage]) throws {
-            let systemPrompt = try AICodeGenRequestBody_V0.getSystemPrompt()
-            
-            let assistantPrompt = """
-                Modify SwiftUI source code based on the request from a user prompt. Use code returned from the last function caller. Adhere to previously defined rules regarding patch and layer behavior in Stitch.
-
-                Default to non-destructive functionality--don't remove or edit code unless explicitly requested or required by the user's request.
-                
-                Adhere to the guidelines specified in this document:
-                
-                \(try StitchAIManager.aiCodeGenSystemPromptGenerator())
-                """
-            
-            self.messages = [.init(role: .system,
-                                   content: systemPrompt),
-                             .init(role: .system,
-                                   content: assistantPrompt)] +
-            toolMessages + [
-                .init(role: .user,
-                      content: userPrompt)
-            ]
-        }
-    }
-}
-
 enum AICodeGenRequestBody_V0 {
     static let systemMarkdownLocation = "AIGraphBuilderSystemPrompt_V0"
 
@@ -98,26 +47,7 @@ enum AICodeGenRequestBody_V0 {
                       content: codeGenAssistantPrompt),
                 .init(role: .user,
                       content: inputsString)
-//                .init(role: .assistant,
-//                      content: assistantPrompt,
-//                      tool_calls: [.init(id: "tc1",
-//                                         type: "function",
-//                                         function: .init(name: self.tool_choice.function.name,
-//                                                         arguments: inputsString))]),
-//                    .init(role: .tool,
-//                          content: inputsString,
-//                          tool_call_id: "tc1",
-//                          name: self.tool_choice.function.name
-//                         )
             ]
         }
     }
-}
-
-final class OpenAIFunctionsHandler {
-    // initial messages...
-    
-    // tool choice...
-    
-    //
 }
