+++ Stitch/Graph/StitchAI/GraphPrompting/Model/SchemaVersions/V0/AIPatchBuilder/AIPatchBuilderRequestBody_V0.swift
@@ -10,12 +10,10 @@ import Foundation
 enum AIPatchBuilderRequestBody_V0 {
     // https://platform.openai.com/docs/api-reference/making-requests
     struct AIPatchBuilderRequestBody: StitchAIRequestableFunctionBody {
-//        static let markdownLocation = "AIPatchBuilderSystemPrompt_V0"
         
         let model: String = "o4-mini-2025-04-16"
         let n: Int = 1
         let temperature: Double = 1.0
-//        let response_format = AIPatchBuilderResponseFormat_V0.AIPatchBuilderResponseFormat()
         let messages: [OpenAIMessage]
         let tools = StitchAIRequestBuilder_V0.StitchAIRequestBuilderFunctions.allFunctions
         let tool_choice = StitchAIRequestBuilder_V0.StitchAIRequestBuilderFunctions.patchBuilder.function
@@ -24,20 +22,6 @@ enum AIPatchBuilderRequestBody_V0 {
         init(userPrompt: String,
              layerDataList: [CurrentAIGraphData.LayerData],
              toolMessages: [OpenAIMessage]) throws {
-//            let responseFormat = AIPatchBuilderResponseFormat_V0.AIPatchBuilderResponseFormat()
-//            let structuredOutputs = responseFormat.json_schema.schema
-//            guard let markdownUrl = Bundle.main.url(forResource: Self.markdownLocation,
-//                                                    withExtension: "md") else {
-//                throw StitchAIStreamingError.markdownNotFound
-//            }
-//            
-//            let assistantPrompt = try String(contentsOf: markdownUrl,
-//                                          encoding: .utf8)
-//            let fullSystemPrompt = "\(systemPrompt)\nUse the following structured outputs schema:\n\(try structuredOutputs.encodeToPrintableString())"
-            
-//            let layerData = try layerDataList.encodeToPrintableString()
-            
-            
             let systemPrompt = try AICodeGenRequestBody_V0.getSystemPrompt()
             let assistantPrompt = try StitchAIManager.aiPatchBuilderSystemPromptGenerator()
